requirements.txt
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
DateTime==4.3
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.19
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1
zope.interface==4.7.1
